moved import-metadata, refactored clean() test
