+++ packages/autotools-scripts/src/import-metadata/import-meta.ts
